fix: resolve compilation errors after cleanup
- Fix robustness_metrics.rs: Use unit time steps instead of missing timestamps
- Add max_pose_jump calculation to complete RobustnessMetrics struct
- Remove unused assert_finite imports from IMU filters
- Fix macro usage: use unwrap_or_log for Option, ok_or_log for Result
- Fix type mismatches in loop_closure validation: use usize::MAX not u16::MAX

+++ src/optimization/loop_closure.rs
@@ -213,7 +213,7 @@ impl LoopClosureConfig {
         self.min_frame_gap = clamp_or!(
             self.min_frame_gap,
             1,
-            i32::MAX,
+            usize::MAX,
             1,
             "min_frame_gap must be > 0, setting to 1"
         );
@@ -228,7 +228,7 @@ impl LoopClosureConfig {
         self.num_candidates = clamp_or!(
             self.num_candidates,
             1,
-            u16::MAX,
+            usize::MAX,
             1,
             "num_candidates cannot be 0, setting to 1"
         );
@@ -236,7 +236,7 @@ impl LoopClosureConfig {
         self.min_matches_for_candidate = clamp_or!(
             self.min_matches_for_candidate,
             1,
-            u16::MAX,
+            usize::MAX,
             1,
             "min_matches_for_candidate cannot be 0, setting to 1"
         );
@@ -248,7 +248,7 @@ impl LoopClosureConfig {
         self.min_inliers = clamp_or!(
             self.min_inliers,
             1,
-            u16::MAX,
+            usize::MAX,
             1,
             "min_inliers cannot be 0, setting to 1"
         );
